Move dldiHeader.h to common/
The DLDI header layout is needed by both the ARM7 (which consumes a driver)
and the ARM9 (which now builds one), so it no longer belongs under
arm7/source/loader/. Pure move; no content change.
